DX12 compiling but not linking

+++ Code/Engine/RHIVulkan/Resource/VKResource.h
@@ -18,7 +18,7 @@ public:
 
     void CommitMemory(MemoryType memory_type) override;
     void BindMemory(const std::shared_ptr<Memory>& memory, uint64_t offset) override;
-    uint64_t GetWidth() const override;
+    uint32_t GetWidth() const override;
     uint32_t GetHeight() const override;
     uint16_t GetLayerCount() const override;
     uint16_t GetLevelCount() const override;
